// vokko-e-voting/src/api/model/event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vokko
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IVoting } from './ivoting';
/**
 * 
 * @export
 * @interface Event
 */
export interface Event {
    /**
     * 
     * @type {string}
     * @memberof Event
     */
    id?: string;
    /**
     * 
     * @type {Date}
     * @memberof Event
     */
    eventDateAndTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof Event
     */
    title?: string | null;
    /**
     * 
     * @type {Array<IVoting>}
     * @memberof Event
     */
    motions?: Array<IVoting> | null;
    /**
     * 
     * @type {Date}
     * @memberof Event
     */
    endDateAndTime?: Date | null;
}